Fixed breaking bug

+++ src/loadgenerator/load_shapes/stable_rampload.py
@@ -11,7 +11,7 @@ WAIT_TIME = int(getenv("LOCUST_WAIT_TIME", "30")) # seconds
 RAMP_DURATION = float(getenv("LOCUST_RAMP_DURATION", "5.0")) # seconds
 STABLE_TAIL = int(getenv("LOCUST_STABLE_P99", "25")) # ms
 VARIANCE_WINDOW = int(getenv("LOCUST_VARIANCE_WINDOW", "30"))
-MAX_VARIANCE = int(getenv("LOCUST_MAX_VARIANCE", "0.3"))
+MAX_VARIANCE = float(getenv("LOCUST_MAX_VARIANCE", "0.3"))
 
 class StableRampLoad(LoadTestShape):
     """
